Rename opts to engineopts
