+++ pkg/frontend/querymiddleware/querysharding_test_utils_test.go
@@ -29,7 +29,7 @@ import (
 	"github.com/grafana/mimir/pkg/storage/series"
 	"github.com/grafana/mimir/pkg/storage/sharding"
 	"github.com/grafana/mimir/pkg/streamingpromql"
-	"github.com/grafana/mimir/pkg/streamingpromql/opts"
+	"github.com/grafana/mimir/pkg/streamingpromql/engineopts"
 	"github.com/grafana/mimir/pkg/util/test"
 )
 
@@ -306,16 +306,16 @@ func TestNewMockShardedQueryable(t *testing.T) {
 	}
 }
 
-type engineOpt func(o *opts.EngineOpts)
+type engineOpt func(o *engineopts.EngineOpts)
 
 func withTimeout(timeout time.Duration) engineOpt {
-	return func(o *opts.EngineOpts) {
+	return func(o *engineopts.EngineOpts) {
 		o.CommonOpts.Timeout = timeout
 	}
 }
 
 func withMaxSamples(samples int) engineOpt {
-	return func(o *opts.EngineOpts) {
+	return func(o *engineopts.EngineOpts) {
 		o.CommonOpts.MaxSamples = samples
 	}
 }
@@ -323,7 +323,7 @@ func withMaxSamples(samples int) engineOpt {
 func newEngineForTesting(t *testing.T, engine string, opts ...engineOpt) (promql.EngineOpts, promql.QueryEngine) {
 	t.Helper()
 
-	mqeOpts := opts.NewTestEngineOpts()
+	mqeOpts := engineopts.NewTestEngineOpts()
 	for _, o := range opts {
 		o(&mqeOpts)
 	}
