+++ pkg/querier/engine/config.go
@@ -11,9 +11,9 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/prometheus/promql"
 
-	"github.com/grafana/mimir/pkg/streamingpromql/opts" //lint:ignore faillint opts is fine
-	"github.com/grafana/mimir/pkg/util/activitytracker" //lint:ignore faillint activitytracker is fine
-	util_log "github.com/grafana/mimir/pkg/util/log"    //lint:ignore faillint log is fine
+	"github.com/grafana/mimir/pkg/streamingpromql/engineopts" //lint:ignore faillint engineopts is fine
+	"github.com/grafana/mimir/pkg/util/activitytracker"       //lint:ignore faillint activitytracker is fine
+	util_log "github.com/grafana/mimir/pkg/util/log"          //lint:ignore faillint log is fine
 )
 
 // Config holds the PromQL engine config exposed by Mimir.
@@ -31,7 +31,7 @@ type Config struct {
 	// series is considered stale.
 	LookbackDelta time.Duration `yaml:"lookback_delta" category:"advanced"`
 
-	MimirQueryEngine opts.EngineOpts `yaml:"mimir_query_engine" category:"experimental"`
+	MimirQueryEngine engineopts.EngineOpts `yaml:"mimir_query_engine" category:"experimental"`
 }
 
 func (cfg *Config) RegisterFlags(f *flag.FlagSet) {
@@ -53,7 +53,7 @@ func (cfg *Config) RegisterFlags(f *flag.FlagSet) {
 }
 
 // NewPromQLEngineOptions returns the PromQL engine options based on the provided config.
-func NewPromQLEngineOptions(cfg Config, activityTracker *activitytracker.ActivityTracker, logger log.Logger, reg prometheus.Registerer) (promql.EngineOpts, opts.EngineOpts) {
+func NewPromQLEngineOptions(cfg Config, activityTracker *activitytracker.ActivityTracker, logger log.Logger, reg prometheus.Registerer) (promql.EngineOpts, engineopts.EngineOpts) {
 	commonOpts := promql.EngineOpts{
 		Logger:               util_log.SlogFromGoKit(logger),
 		Reg:                  reg,
