+++ pkg/streamingpromql/benchmarks/comparison_test.go
@@ -31,7 +31,7 @@ import (
 	"github.com/grafana/mimir/pkg/querier"
 	"github.com/grafana/mimir/pkg/querier/stats"
 	"github.com/grafana/mimir/pkg/streamingpromql"
-	"github.com/grafana/mimir/pkg/streamingpromql/opts"
+	"github.com/grafana/mimir/pkg/streamingpromql/engineopts"
 	"github.com/grafana/mimir/pkg/streamingpromql/testutils"
 	"github.com/grafana/mimir/pkg/util/validation"
 )
@@ -42,7 +42,7 @@ func BenchmarkQuery(b *testing.B) {
 	q := createBenchmarkQueryable(b, MetricSizes)
 	cases := TestCases(MetricSizes)
 
-	opts := opts.NewTestEngineOpts()
+	opts := engineopts.NewTestEngineOpts()
 	prometheusEngine := promql.NewEngine(opts.CommonOpts)
 	mimirEngine, err := streamingpromql.NewEngine(opts, streamingpromql.NewStaticQueryLimitsProvider(0), stats.NewQueryMetrics(nil), streamingpromql.NewQueryPlanner(opts), log.NewNopLogger())
 	require.NoError(b, err)
@@ -94,7 +94,7 @@ func TestBothEnginesReturnSameResultsForBenchmarkQueries(t *testing.T) {
 	q := createBenchmarkQueryable(t, metricSizes)
 	cases := TestCases(metricSizes)
 
-	opts := opts.NewTestEngineOpts()
+	opts := engineopts.NewTestEngineOpts()
 	prometheusEngine := promql.NewEngine(opts.CommonOpts)
 	limitsProvider := streamingpromql.NewStaticQueryLimitsProvider(0)
 	queryMetrics := stats.NewQueryMetrics(nil)
@@ -124,7 +124,7 @@ func TestBothEnginesReturnSameResultsForBenchmarkQueries(t *testing.T) {
 func TestBenchmarkSetup(t *testing.T) {
 	q := createBenchmarkQueryable(t, []int{1})
 
-	opts := opts.NewTestEngineOpts()
+	opts := engineopts.NewTestEngineOpts()
 	mimirEngine, err := streamingpromql.NewEngine(opts, streamingpromql.NewStaticQueryLimitsProvider(0), stats.NewQueryMetrics(nil), streamingpromql.NewQueryPlanner(opts), log.NewNopLogger())
 	require.NoError(t, err)
