+++ pkg/streamingpromql/engineopts/config.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
-package opts
+package engineopts
 
 import (
 	"flag"
