+++ pkg/streamingpromql/functions_test.go
@@ -14,8 +14,8 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/grafana/mimir/pkg/querier/stats"
+	"github.com/grafana/mimir/pkg/streamingpromql/engineopts"
 	"github.com/grafana/mimir/pkg/streamingpromql/operators/functions"
-	"github.com/grafana/mimir/pkg/streamingpromql/opts"
 )
 
 // This test ensures that all functions correctly merge series after dropping the metric name.
@@ -29,7 +29,7 @@ func TestFunctionDeduplicateAndMerge(t *testing.T) {
 	`
 
 	storage := promqltest.LoadedStorage(t, data)
-	opts := opts.NewTestEngineOpts()
+	opts := engineopts.NewTestEngineOpts()
 	engine, err := NewEngine(opts, NewStaticQueryLimitsProvider(0), stats.NewQueryMetrics(nil), NewQueryPlanner(opts), log.NewNopLogger())
 	require.NoError(t, err)
