+++ pkg/streamingpromql/optimize/plan/skip_histogram_decoding_test.go
@@ -10,9 +10,9 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/grafana/mimir/pkg/streamingpromql"
+	"github.com/grafana/mimir/pkg/streamingpromql/engineopts"
 	"github.com/grafana/mimir/pkg/streamingpromql/optimize/plan"
 	"github.com/grafana/mimir/pkg/streamingpromql/optimize/plan/commonsubexpressionelimination"
-	"github.com/grafana/mimir/pkg/streamingpromql/opts"
 	"github.com/grafana/mimir/pkg/streamingpromql/testutils"
 	"github.com/grafana/mimir/pkg/streamingpromql/types"
 )
@@ -371,7 +371,7 @@ func TestSkipHistogramDecodingOptimizationPass(t *testing.T) {
 	timeRange := types.NewInstantQueryTimeRange(time.Now())
 	observer := streamingpromql.NoopPlanningObserver{}
 
-	opts := opts.NewTestEngineOpts()
+	opts := engineopts.NewTestEngineOpts()
 	planner := streamingpromql.NewQueryPlannerWithoutOptimizationPasses(opts)
 	planner.RegisterQueryPlanOptimizationPass(commonsubexpressionelimination.NewOptimizationPass(true, nil))
 	planner.RegisterQueryPlanOptimizationPass(plan.NewSkipHistogramDecodingOptimizationPass())
